requirements.txt
fastapi
uvicorn
python-multipart
cairosvg
pdf2image
Pillow
